% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_efficiency.R
\name{analyze_efficiency}
\alias{analyze_efficiency}
\title{Calculate PCR efficiency over multiple amplification models. This function is a wrapper around qpcR:expfit()}
\usage{
analyze_efficiency(data, method = "outlier", progress = TRUE, ...)
}
\arguments{
\item{data}{A list of models built with prepare_batch() or read_[EQUIPMENT]()}

\item{progress}{Should a progress bar be shown?}

\item{...}{Other arguments passed to qpcR::expfit()}
}
\value{
A data frame with parameters from qpcR::expfit()
}
\description{
Calculate PCR efficiency over multiple amplification models. This function is a wrapper around qpcR:expfit()
}
