% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_models.R
\name{analyze_models}
\alias{analyze_models}
\title{Analyze models from model_qpcr() using efficiency() from the qpcR-package}
\usage{
analyze_models(models, progress = TRUE, ...)
}
\arguments{
\item{models}{a list of models created with qpcrpal::model_qpcr()}

\item{progress}{Should a progress bar be shown?}

\item{...}{Other arguments passed to qpcr::efficiency()}

\item{sampleID}{Col numbers corresponding to ID, condition, timepoint and replicate}
}
\description{
Analyze models from model_qpcr() using efficiency() from the qpcR-package
}
