% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPrep.R
\name{batchPrep}
\alias{batchPrep}
\title{Import and prepare a batch for analysis}
\usage{
batchPrep(folder, ..., progress = TRUE, equipment = "ABI")
}
\arguments{
\item{folder}{A path to a folder containing multiple exports from ABI7500}

\item{...}{Argumnts passed to read_[EQUIPMENT]}

\item{progress}{logical, if TRUE a progressbar is shown works only on windows}

\item{equipment}{Specify from what equipment the export originates}
}
\description{
Import and prepare a batch for analysis
}
