% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_qpcr.R
\name{model_qpcr}
\alias{model_qpcr}
\title{Fit a model to batch of raw data using qpcR::pcrfit}
\usage{
model_qpcr(rawData, sampleID = c(2, 3, 4, 5), replicate = TRUE,
  progress = TRUE, ...)
}
\arguments{
\item{rawData}{A data frame compiled with read_*() or batchPrep()}

\item{sampleID}{Col numbers corresponding to ID, condition, timepoint and replicate}

\item{replicate}{Logical value indicating if replicates should be modeled togheter}

\item{progress}{Should a progress bar be shown?}

\item{...}{Other arguments passed to pcrfit()}
}
\description{
Fit a model to batch of raw data using qpcR::pcrfit
}
