% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_batch.R
\name{prepare_batch}
\alias{prepare_batch}
\title{Import and prepare a batch for analysis, this function replaces batchPrep()}
\usage{
prepare_batch(folder, ..., progress = TRUE, equipment = "ABI")
}
\arguments{
\item{folder}{A path to a folder containing multiple exports from ABI7500 or Quantstudio 5}

\item{...}{Argumnts passed to read_[EQUIPMENT]}

\item{progress}{logical, if TRUE a progressbar is shown works only on windows}

\item{equipment}{Specify from what equipment the export originates, "ABI" for ABI7500, "quant" for quantstudio 5.}
}
\description{
Import and prepare a batch for analysis, this function replaces batchPrep()
}
