% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quant5.R
\name{read_quant5}
\alias{read_quant5}
\title{Extract raw data (normalized reporter), sample name and target name from Quantstudio 5 exports.}
\usage{
read_quant5(filename, sample_separator = " ", sample_ID = 1,
  sample_condition = 2, sample_timepoint = 3, sample_replicate = 4,
  start_cycle = 1, end_cycle = 40, skip = 19)
}
\arguments{
\item{filename}{A path or filename of an excel file containing 'Results' and 'Raw Data' exported from QuantStudio}

\item{sample_separator}{A character string specifying separator of 'Sample Name' string.}

\item{sample_ID}{A integer specifying sample sub-info position in 'Sample Name' from export. Default=1}

\item{sample_condition}{Default=2}

\item{sample_timepoint}{Default=3}

\item{sample_replicate}{Default=4}

\item{start_cycle}{First cycle to extract from export, default=1}

\item{end_cycle}{Last cycle to extract from export, default=40}

\item{skip}{How many rows needs to be skipped before column names in the export?}
}
\value{
A data frame containing sample and target information and raw data over 40 cycles.
}
\description{
Extract raw data (normalized reporter), sample name and target name from Quantstudio 5 exports.
}
