% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_models.R
\name{test_models}
\alias{test_models}
\title{Test model fit on models prepared with qpcrpal::model_qpcr(), based on qpcR::mselect().}
\usage{
test_models(models, targetID = 4)
}
\arguments{
\item{models}{a list of models from model_qpcr()}

\item{targetID}{Col numbers corresponding to target identifier}

\item{progress}{Should a progress bar be shown?}
}
\value{
A data frame with results from mselect() tests of model fit based on AICc, lower is better.
}
\description{
Test model fit on models prepared with qpcrpal::model_qpcr(), based on qpcR::mselect().
}
